Fix unsigned vs signed comparisons and 32 bit string formats
Fix unsigned vs signed comparison warnings.

For 64 bit numbers printed in gRPC, the string formats assume that you
are running on a 64 bit machine.  Use inttypes.h to make it portable.
Also, use size_t format strings for the same reason.

+++ src/core/ext/filters/load_reporting/server_load_reporting_filter.cc
@@ -200,8 +200,9 @@ void ServerLoadReportingCallData::StoreClientIpAndLrToken(const char* lr_token,
   if (lr_token_len != 0) {
     strncpy(cur_pos, lr_token, lr_token_len);
   }
-  GPR_ASSERT(cur_pos + lr_token_len - client_ip_and_lr_token_ ==
-             long(client_ip_and_lr_token_len_));
+  GPR_ASSERT(
+      static_cast<size_t>(cur_pos + lr_token_len - client_ip_and_lr_token_) ==
+      client_ip_and_lr_token_len_);
 }
 
 grpc_filtered_mdelem ServerLoadReportingCallData::RecvInitialMetadataFilter(
